Added an option to child_process.spawn
child_process.spawn now accepts an array of group ids (or group names),
in addition to the numerical group id or group name.

The signature for the internal method ChildProcess::Spawn was also changed.
It now receives an user id and an array of group ids, instead of a
single user id/name and a single group id/name.